rename and move set_cb_label to separate file
