+++ external_parser/log_converter.h
@@ -13,6 +13,6 @@
 namespace v2 = reinforcement_learning::messages::flatbuff::v2;
 
 namespace log_converter {
-void build_cb_json(std::ofstream &outfile, const joined_event &je, float reward,
+void build_cb_json(std::ofstream &outfile, const joined_event::joined_event &je, float reward,
                    float original_reward);
 } // namespace log_converter
